fix: Set `enable_domain_permissions_policy` to `false` by default
The changes in this commit update the default value of the `enable_domain_permissions_policy` variable in the `variables.tf` file from `true` to `false`. This change is made to better reflect the desired default behavior for the domain permissions policy.

Additionally, the commit includes an update to the `.terraform.lock.hcl` file, which is a lock file that ensures the same provider versions are used across different environments.